requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Mail==0.10.0
WTForms==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.0.0
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
